Fix memory management helpers

+++ tile/kernel/memory.h
@@ -19,7 +19,7 @@ const extern uint32_t* data_end;
 const extern uint32_t* bss_begin;
 const extern uint32_t* bss_end;
 
-extern uint64_t virt_to_phys(uint32_t x);
+extern uint32_t virt_to_phys(uint32_t x);
 extern uint32_t phys_to_virt(uint32_t x);
 
 extern struct memory_map memory_map;
